+++ src/main/java/com/meritoki/retina/application/desktop/model/project/Point.java
@@ -10,8 +10,8 @@ import org.codehaus.jackson.map.ObjectWriter;
 
 public class Point {
 	
-	public int x = 0;
-	public int y = 0;
+	public double x = 0;
+	public double y = 0;
 	
     @JsonIgnore
     @Override
